Extract AndroidAppView
