+++ game-of-life-multiplatform/android/src/main/res/layout/activity_game_of_life.xml
@@ -1,35 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:tools="http://schemas.android.com/tools"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  android:orientation="vertical"
   tools:context="com.novoda.gol.GameOfLife">
 
-  <LinearLayout
+  <com.novoda.gol.presentation.AndroidAppView
+    android:id="@+id/appView"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <Button
-      android:id="@+id/controlButton"
-      android:layout_width="0dp"
-      android:layout_weight="1"
-      android:layout_height="wrap_content" />
-
-    <Button
-      android:id="@+id/patternSelectionButton"
-      android:text="Choose Pattern"
-      android:layout_width="0dp"
-      android:layout_weight="1"
-      android:layout_height="wrap_content" />
-
-  </LinearLayout>
-
-
-  <com.novoda.gol.presentation.AndroidBoardView
-    android:id="@+id/boardView"
-    android:layout_width="wrap_content"
     android:layout_height="wrap_content" />
 
-
-</LinearLayout>
+</FrameLayout>
